# api/openapi.yaml
openapi: 3.0.3
info:
  title: Warta API
  version: 1.0.0
  description: |
    REST API untuk platform artikel Warta: akun dan role, kategori, tag,
    artikel, dan komentar.

    **Autentikasi.** Login atau register menghasilkan `access_token` (JWT,
    berlaku singkat) dan `refresh_token` (berlaku lebih lama, sekali pakai).
    Kirim access token di header `Authorization: Bearer <token>`. Saat access
    token kedaluwarsa, tukar refresh token di `/auth/refresh` untuk mendapat
    pasangan baru. Refresh token lama langsung tidak berlaku; bila dipakai lagi,
    semua sesi akun itu dicabut.

    Endpoint publik tetap bisa dipanggil tanpa token. Namun token yang dikirim
    tapi tidak valid atau kedaluwarsa selalu ditolak dengan 401.

    **Role.** Akun baru selalu `reader` (bisa berkomentar). Admin menaikkannya
    menjadi `author` (bisa menulis artikel) atau `admin`.

    **Verifikasi email.** Register mengirim email berisi tautan
    `APP_URL/verify-email?token=...`; frontend meneruskan token itu ke
    `/auth/verify-email`. Bila `REQUIRE_EMAIL_VERIFICATION` aktif (bawaan di
    production), akun yang belum terverifikasi ditolak saat berkomentar atau
    melapor dengan kode `email_not_verified`. Lupa password bekerja serupa
    lewat `APP_URL/reset-password?token=...`.

    **Bentuk response.** Sukses: `{"data": ...}`, ditambah `meta` untuk daftar
    berhalaman. Gagal: `{"error": {"code", "message", "fields"}}`.
servers:
  - url: /
...

...
security:
  - {}
  - bearerAuth: []
...
paths:
...
  /sitemap.xml:
    get:
      operationId: sitemap
      tags: [SEO]
      summary: Sitemap beranda, kategori, dan artikel terbit
      description: Semua tautan menuju frontend (`APP_URL`). Di-cache 15 menit.
      security: [{}]
      responses:
        "200":
          description: OK
          content:
            application/xml:
              schema: { type: string }
  /feed.xml:
    get:
      operationId: rssFeed
      tags: [SEO]
      summary: RSS 2.0 berisi 20 artikel terbit terbaru
      security: [{}]
      responses:
        "200":
          description: OK
          content:
            application/rss+xml:
              schema: { type: string }
...
  /api/v1/me/feed:
    get:
      tags: [Penulis]
      operationId: listMyFeed
      summary: Artikel terbit dari penulis yang diikuti
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Sort"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArticleListEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/articles:
    get:
      operationId: listArticles
      tags: [Artikel]
      summary: Daftar artikel
      description: |
        Tanpa `status`, hanya artikel yang sudah terbit yang tampil. Status lain
        (`draft`, `archived`, `all`) hanya untuk admin. Daftar berisi cuplikan
        (`excerpt`), bukan isi lengkap.
      security: [{}, { bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/CategoryFilter"
        - $ref: "#/components/parameters/TagFilter"
        - name: author
          in: query
          description: Id penulis
          schema: { type: integer, format: int64 }
        - name: status
          in: query
          schema: { type: string, enum: [draft, scheduled, published, archived, all] }
        - $ref: "#/components/parameters/Sort"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArticleListEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createArticle
      tags: [Artikel]
      summary: Menulis artikel (author atau admin)
      description: |
        Slug dibuat dari judul dan dibuat unik. Tag yang belum ada dibuat
        otomatis; nama tag dirapikan menjadi huruf kecil.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ArticleRequest" }
      responses:
        "201":
          description: Dibuat
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArticleEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
  /api/v1/articles/{id}/revisions:
    get:
      tags: [Artikel]
      operationId: listRevisions
      summary: Riwayat revisi, terbaru dulu (penulis artikel dan admin)
      description: |
        Revisi baru tercatat setiap kali judul, isi, kategori, tag, atau sampul
        berubah; perubahan status saja tidak. Paling banyak 50 revisi terbaru
        disimpan per artikel.
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/RevisionSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/articles/{id}/revisions/{revision}:
    get:
      tags: [Artikel]
      operationId: getRevision
      summary: Isi lengkap satu revisi
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/Revision"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Revision" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/articles/{id}/revisions/{revision}/restore:
    post:
      tags: [Artikel]
      operationId: restoreRevision
      summary: Memulihkan artikel ke isi sebuah revisi
      description: |
        Judul, isi, kategori, tag, dan sampul dikembalikan; status tidak
        berubah. Pemulihan tercatat sebagai revisi baru, jadi bisa dibatalkan.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/Revision"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArticleEnvelope" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PerPage:
      name: per_page
      in: query
      description: Nilai di atas `MAX_PER_PAGE` dipangkas.
      schema: { type: integer, minimum: 1, default: 10 }
    Search:
      name: q
      in: query
      description: |
        Cari di judul dan isi, tidak peka huruf besar. Semua kata harus ada,
        dan kata yang belum lengkap juga cocok ("golan" menemukan "golang").
        Hasil membawa field snippet berisi potongan isi di sekitar kata itu.
      schema: { type: string }
    CategoryFilter:
      name: category
      in: query
      description: Slug kategori
      schema: { type: string }
    TagFilter:
      name: tag
      in: query
      description: Slug tag
      schema: { type: string }
    Sort:
      name: sort
      in: query
      description: |
        popular menimbang jumlah dibaca, suka (x5), dan komentar (x3).
        relevance (bawaan bila ada `q`) mengurutkan menurut kecocokan kata
        kunci, dengan kecocokan di judul berbobot tiga kali.
      schema: { type: string, enum: [relevance, newest, oldest, title, updated, popular] }
    Revision:
      name: revision
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
  responses:
    BadRequest:
      description: Parameter atau body tidak bisa dibaca
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Token tidak ada, tidak valid, atau kedaluwarsa
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Role atau kepemilikan tidak mengizinkan aksi ini
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Tidak ditemukan
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: Bentrok dengan data yang disimpan bersamaan
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    PayloadTooLarge:
      description: Body lebih dari 1 MB
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ValidationFailed:
      description: Validasi gagal, rinciannya per field di `error.fields`
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          example:
            error:
              code: validation_failed
              message: validasi gagal
              fields:
                title: title minimal 20 karakter
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - bad_request
                - unauthorized
                - forbidden
                - email_not_verified
                - not_found
                - method_not_allowed
                - conflict
                - payload_too_large
                - validation_failed
                - too_many_requests
                - internal_error
            message: { type: string }
            fields:
              type: object
              additionalProperties: { type: string }
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        per_page: { type: integer }
        total: { type: integer, format: int64 }
        total_pages: { type: integer }
...
    Author:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        avatar_url: { type: string, nullable: true }
...
    RevisionSummary:
      type: object
      properties:
        id: { type: integer, format: int64 }
        title: { type: string }
        editor:
          type: object
          nullable: true
          description: null bila akun penyunting sudah dihapus
          properties:
            id: { type: integer, format: int64 }
            name: { type: string }
        characters: { type: integer, description: Panjang isi }
        created_at: { type: string, format: date-time }
    Revision:
      allOf:
        - $ref: "#/components/schemas/RevisionSummary"
        - type: object
          properties:
            content: { type: string }
            category_id: { type: integer, format: int64, nullable: true, description: null bila kategorinya sudah dihapus }
            tags:
              type: array
              items: { type: string }
            cover_image: { type: string, nullable: true }
...
    CategoryRef:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        slug: { type: string }
...
    TagRef:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        slug: { type: string }
    ArticleStatus:
      type: string
      description: scheduled terbit otomatis pada scheduled_at; sampai saat itu hanya terlihat oleh penulis dan admin.
      enum: [draft, scheduled, published, archived]
    ArticleSummary:
      type: object
      properties:
        id: { type: integer, format: int64 }
        title: { type: string }
        slug: { type: string }
        excerpt: { type: string }
        status: { $ref: "#/components/schemas/ArticleStatus" }
        author: { $ref: "#/components/schemas/Author" }
        category: { $ref: "#/components/schemas/CategoryRef" }
        tags:
          type: array
          items: { $ref: "#/components/schemas/TagRef" }
        cover_image:
          type: string
          nullable: true
          description: Path gambar sampul, misalnya /uploads/3f2a....webp
        comment_count: { type: integer }
        like_count: { type: integer }
        view_count: { type: integer }
        reading_minutes: { type: integer, description: Perkiraan lama membaca }
        published_at:
          type: string
          format: date-time
          nullable: true
          description: Waktu pertama kali terbit
        scheduled_at:
          type: string
          format: date-time
          nullable: true
          description: Waktu terbit otomatis, hanya untuk status scheduled
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        snippet:
          type: string
          description: Hanya di hasil pencarian (q), potongan isi di sekitar kata yang dicari
    Article:
      allOf:
        - $ref: "#/components/schemas/ArticleSummary"
        - type: object
          properties:
            content: { type: string, description: Isi dalam format Markdown }
            liked: { type: boolean, description: Selalu false bila belum login }
            bookmarked: { type: boolean, description: Selalu false bila belum login }
...
    ArticleRequest:
      type: object
      required: [title, content, category_id, status]
      properties:
        title: { type: string, minLength: 20, maxLength: 200 }
        content: { type: string, minLength: 200, maxLength: 100000, description: Markdown }
        category_id: { type: integer, format: int64 }
        cover_image:
          type: string
          description: Path hasil POST /api/v1/uploads. Kosong berarti tanpa sampul.
        tags:
          type: array
          maxItems: 10
          items: { type: string, minLength: 2, maxLength: 50 }
        status: { $ref: "#/components/schemas/ArticleStatus" }
        scheduled_at:
          type: string
          format: date-time
          description: Wajib untuk status scheduled, di masa depan dan paling jauh setahun
      example:
        title: Panduan Membangun REST API dengan Golang
        content: >-
          Artikel ini membahas cara menyusun REST API dengan Go dari nol: menata
          struktur folder, memisahkan handler, service, dan repository, menulis
          migrasi database, sampai menguji endpoint secara end-to-end dengan MySQL.
        category_id: 1
        tags: [golang, backend]
        status: draft
...
    ArticleEnvelope:
      type: object
      properties:
        data: { $ref: "#/components/schemas/Article" }
    ArticleListEnvelope:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/ArticleSummary" }
        meta: { $ref: "#/components/schemas/PaginationMeta" }
...